feat: close the review period at REVIEW_DEADLINE
Reviewer OTP sign-in is refused (surfaced as a toast on /otp-login),
GET/POST /api/reviews return 403 (writes blocked for admins too), and the
/reviews pages render a closed card for still-valid sessions. The check runs
per-request — pages are force-dynamic so the answer isn't baked in at build
time — so no redeploy is needed at cutoff. Deadline set to EOD 24 Jul 2026
Africa/Nairobi in wrangler.jsonc; unset means reviews stay open.
