+++ .env.example
@@ -9,6 +9,10 @@ SESSION_EXPIRY_DAYS="3"
 # Call for Speakers — set to "true" to accept submissions. Defaults to closed.
 CALL_FOR_SPEAKERS_OPEN="false"
 
+# ISO timestamp (include the UTC offset) after which reviewer sign-in and
+# review reads/writes are refused. Empty or unset means reviews stay open.
+REVIEW_DEADLINE=""
+
 DATABASE_URL="postgresql://bigzoo:postgres@localhost:5432/gophercon"
 RESEND_API_KEY=""
