--- .eslintrc.build.json
@@ -1,6 +1,0 @@
-{
-  "extends": [],
-  "rules": {
-    "*": "off"
-  }
-}
