+++ src/actions/auth/otp.ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 import { db } from '@/src/db';
 import { Resend } from 'resend';
 import { generateOtp, hashOtp } from '@/src/lib/otp';
-import { OTP_EXPIRY_MINUTES } from '@/src/lib/config';
+import { OTP_EXPIRY_MINUTES, isReviewPeriodOpen } from '@/src/lib/config';
 import { getEmailRole } from '@/src/lib/authorizedUsers';
 import { OtpFormState } from '@/src/types/otp';
 import { OtpEmail } from '@/src/notification/email/templates/otp-verification';
@@ -28,6 +28,10 @@ export async function sendOtp(_: OtpFormState, formData: FormData): Promise<OtpF
   const role = await getEmailRole(normalizedEmail);
   if (!role) {
     return { errors: { email: ['Email not authorized'] } };
+  }
+
+  if (role === 'reviewer' && !isReviewPeriodOpen()) {
+    return { errors: { _form: ['The review period is now closed. Thank you for reviewing!'] } };
   }
 
   try {
