+++ src/lib/config.ts
@@ -26,3 +26,21 @@ export const SESSION_EXPIRY_DAYS = parsePositiveInt(process.env.SESSION_EXPIRY_D
 // set CALL_FOR_SPEAKERS_OPEN=true to reopen.
 export const CALL_FOR_SPEAKERS_OPEN = parseBoolean(process.env.CALL_FOR_SPEAKERS_OPEN, false);
 
+const parseDate = (value: string | undefined): Date | null => {
+  if (!value || value.trim() === '') {
+    return null;
+  }
+  const parsed = new Date(value.trim());
+  return Number.isNaN(parsed.getTime()) ? null : parsed;
+};
+
+// ISO timestamp (with explicit UTC offset) after which reviewer sign-in and
+// review reads/writes are refused. Unset or invalid means reviews stay open.
+export const REVIEW_DEADLINE = parseDate(process.env.REVIEW_DEADLINE);
+
+// A function, not a module-scope boolean: Workers isolates are long-lived, so
+// a value computed at cold start would never cross the deadline.
+export function isReviewPeriodOpen(): boolean {
+  return REVIEW_DEADLINE === null || new Date() < REVIEW_DEADLINE;
+}
+
